refactor: add CardViewerViewModel
it is a extraction of the general card viewing functionality so it can be reused in the reviewer and template previewer
